// src/redux/types/carCardTypes.ts
export enum CarCardActionTypes {
  SET_SELECTED_CAR = 'SET_SELECTED_CAR',
  RESET_CAR = 'RESET_CAR',
  UPDATE_CARD = 'UPDATE_CARD',
  PUT_CAR_FETCHING = 'PUT_CAR_FETCHING',
  PUT_CAR_SUCCESS = 'PUT_CAR_SUCCESS',
  PUT_CAR_ERROR = 'PUT_CAR_ERROR',
  POST_CAR_FETCHING = 'POST_CAR_FETCHING',
  POST_CAR_SUCCESS = 'POST_CAR_SUCCESS',
  POST_CAR_ERROR = 'POST_CAR_ERROR',
  DELETE_CAR_BY_ID_FETCHING = 'DELETE_CAR_BY_ID_FETCHING',
  DELETE_CAR_BY_ID_SUCCESS = 'DELETE_CAR_BY_ID_SUCCESS',
  DELETE_CAR_BY_ID_ERROR = 'DELETE_CAR_BY_ID_ERROR',
}